Move PermissionIcons component to src/components/collectionManagement
