+++ app/packs/src/components/collectionManagement/PermissionIcons.js
